requirements.txt
flask==3.1.0
flask-cors==5.0.1
numpy==2.2.3
pandas==2.2.3
scikit-learn==1.6.1
Pillow==11.1.0
pytesseract==0.3.13
joblib==1.4.2
gunicorn==23.0.0
